:sparkles: Desafio: conhecendo o BigDecimal

+++ alura-cursos/ano-2024/kotlin/src/main/ferreira/lks/alugames/models/PlanoAssinatura.kt
@@ -1,5 +1,7 @@
 package lks.alugames.models
 
+import java.math.BigDecimal
+
 class PlanoAssinatura(
     tipo: String,
     val mensalidade: Double,
@@ -7,12 +9,12 @@ class PlanoAssinatura(
     private val porcentagemDesconto: Int
 ) : Plano(tipo) {
 
-    override fun getValorAluguel(aluguel: Aluguel): Double {
+    override fun getValorAluguel(aluguel: Aluguel): BigDecimal {
         return if (totalJogosInclusos > aluguel.gamer.getJogosDoMes(aluguel.periodo.dataInicial.monthValue).size) {
-            0.0
+            0.0.toBigDecimal()
         } else {
             val valorOriginal = super.getValorAluguel(aluguel)
-            return if (aluguel.gamer.media > 8) valorOriginal - valorOriginal * porcentagemDesconto / 100 else valorOriginal
+            return if (aluguel.gamer.media > 8) valorOriginal - valorOriginal * (porcentagemDesconto / 100).toBigDecimal() else valorOriginal
         }
     }
 }
